simplify the API; leverage AndValue for additional assertion chaining
