+++ src/Functional.Primitives.FluentAssertions/Functional.Primitives.FluentAssertions/Extensions/ResultExtensions.cs
@@ -4,7 +4,6 @@ namespace Functional.Primitives.FluentAssertions.Extensions
 {
 	internal static class ResultExtensions
 	{
-		public static bool IsSuccessful<TSuccess, TFailure>(this Result<TSuccess, TFailure> source) => source.Match(_ => true, _ => false);
 		public static TSuccess GetSuccessValue<TSuccess, TFailure>(this Result<TSuccess, TFailure> source) => source.Match(value => value, _ => throw new InvalidOperationException("Attempted to access value from success case, but none exists!"));
 		public static TFailure GetFailureValue<TSuccess, TFailure>(this Result<TSuccess, TFailure> source) => source.Match(_ => throw new InvalidOperationException("Attempted to access value from failure case, but none exists!"), value => value);
 	}
